Lykke.Service.OAuth/src/Lykke.Service.OAuth/Attributes/FileTypesAttribute.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.IO;
using System.Linq;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;

namespace WebAuth.Attributes
{
    [AttributeUsage(AttributeTargets.Property)]
    public class FileTypesAttribute : ValidationAttribute, IClientModelValidator
    {
        private readonly List<string> _types;

        public FileTypesAttribute(string types)
        {
            _types = types.Split(',').ToList();
        }

        public void AddValidation(ClientModelValidationContext context)
        {
            if (context == null)
                throw new ArgumentNullException(nameof(context));

            MergeAttribute(context.Attributes, "data-val", "true");
            MergeAttribute(context.Attributes, "data-val-filetypes", ErrorMessage);
        }

        protected override ValidationResult IsValid(object value, ValidationContext validationContext)
        {
            if (value == null) return ValidationResult.Success;

            var fileExt = Path.GetExtension((value as IFormFile).FileName).Substring(1);

            if (!_types.Contains(fileExt, StringComparer.OrdinalIgnoreCase))
                return new ValidationResult(ErrorMessage);

            return ValidationResult.Success;
        }

        private bool MergeAttribute(IDictionary<string, string> attributes, string key, string value)
        {
            if (attributes.ContainsKey(key))
                return false;

            attributes.Add(key, value);
            return true;
        }
    }
}